// frontend/src/lib/api/generated/models/repoBrowserBlobResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { RepoBrowserBlob } from "./repoBrowserBlob.ts";
import type { RepoBrowserRef } from "./repoBrowserRef.ts";
import type { RepoRefResponse } from "./repoRefResponse.ts";

export interface RepoBrowserBlobResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  blob: RepoBrowserBlob;
  ref: RepoBrowserRef;
  repo: RepoRefResponse;
}