// frontend/src/lib/api/generated/models/repoBrowserCommitResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { RepoBrowserCommit } from "./repoBrowserCommit.ts";
import type { RepoBrowserRef } from "./repoBrowserRef.ts";
import type { RepoRefResponse } from "./repoRefResponse.ts";

export interface RepoBrowserCommitResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  commit: RepoBrowserCommit;
  path: string;
  ref: RepoBrowserRef;
  repo: RepoRefResponse;
}