// frontend/src/lib/api/generated/models/repoBrowserHistoryResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { RepoBrowserCommit } from "./repoBrowserCommit.ts";
import type { RepoBrowserRef } from "./repoBrowserRef.ts";
import type { RepoRefResponse } from "./repoRefResponse.ts";

export interface RepoBrowserHistoryResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  commits: RepoBrowserCommit[];
  path: string;
  ref: RepoBrowserRef;
  repo: RepoRefResponse;
}