// frontend/src/lib/api/generated/models/repoBrowserLastChangedResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { RepoBrowserLastChangedResponseCommits } from "./repoBrowserLastChangedResponseCommits.ts";
import type { RepoBrowserRef } from "./repoBrowserRef.ts";
import type { RepoRefResponse } from "./repoRefResponse.ts";

export interface RepoBrowserLastChangedResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  commits: RepoBrowserLastChangedResponseCommits;
  ref: RepoBrowserRef;
  repo: RepoRefResponse;
}