// frontend/src/lib/api/generated/models/repoBrowserRefsResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { RepoBrowserRef } from "./repoBrowserRef.ts";
import type { RepoRefResponse } from "./repoRefResponse.ts";

export interface RepoBrowserRefsResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  default_ref: RepoBrowserRef;
  refs: RepoBrowserRef[];
  repo: RepoRefResponse;
  truncated: boolean;
}