// frontend/src/lib/api/generated/models/repoBrowserTreeResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { RepoBrowserRef } from "./repoBrowserRef.ts";
import type { RepoBrowserTreeEntry } from "./repoBrowserTreeEntry.ts";
import type { RepoRefResponse } from "./repoRefResponse.ts";

export interface RepoBrowserTreeResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  entries: RepoBrowserTreeEntry[];
  ref: RepoBrowserRef;
  repo: RepoRefResponse;
  truncated: boolean;
}